requirements.txt
Flask
Flask-SQLAlchemy
requests
APScheduler
python-telegram-bot
plotly
